refactor: move flow manager into dedicated module
